backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
openai==2.29.0
pydantic==2.12.5
alembic==1.18.4
